fix deactive feature for user account

+++ app/routers/users.py
@@ -1,7 +1,9 @@
 from fastapi import APIRouter, Depends, File, HTTPException, UploadFile
+from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from app.core.cloudinary import delete_profile_photo, upload_profile_photo
 from app.database import get_db
+from app.models.transaction import Order, TransactionStatus
 from app.models.user import User
 from app.schemas.user import ChangeEmail, ChangePassword, UpdateUsername, UserResponse
 from app.crud import user as crud_user
@@ -133,10 +135,25 @@ async def delete_photo(
 
 @router.delete("/me")
 async def delete_my_account(current_user: User = Depends(get_current_user), db: AsyncSession = Depends(get_db)):
+    result = await db.execute(
+        select(Order)
+        .where(
+            Order.user_id == current_user.id,
+            Order.status == TransactionStatus.PENDING
+        )
+    )
+    pending_orders = result.scalars().all()
+    
+    if pending_orders:
+        raise HTTPException(
+            status_code=400,
+            detail="Masih terdapat transaksi 'PENDING', tidak dapat menghapus akun"
+        )
+    
     if current_user.image:
         await delete_profile_photo(current_user.id)
         
     await crud_user.delete_user(db, current_user)
-    return{
+    return {
         "message": "Account permanently deleted"
     }
